=feat(multimedia): COCO-4479, extract useCameras hook for VideoRecorder
