+++ packages/react/src/modules/multimedia/VideoRecorder/VideoRecorder.tsx
@@ -20,7 +20,7 @@ import {
   Toolbar,
   ToolbarItem,
 } from '../../../components';
-import { useBrowserInfo, useUpload } from '../../../hooks';
+import { useUpload } from '../../../hooks';
 import {
   IconPause,
   IconPlayFilled,
@@ -30,6 +30,7 @@ import {
   IconRefresh,
   IconSave,
 } from '../../icons/components';
+import { useCameras } from './useCameras';
 
 export interface VideoRecorderProps {
   appCode: string;
@@ -39,9 +40,6 @@ export interface VideoRecorderProps {
   onRecordUpdated?: (recordURL?: string) => void;
   hideSaveAction?: boolean;
 }
-
-const VIDEO_HEIGHT = 9;
-const VIDEO_WIDTH = 16;
 
 export interface VideoRecorderRef {
   save: () => Promise<WorkspaceElement | undefined>;
@@ -59,8 +57,15 @@ const VideoRecorder = forwardRef(
     }: VideoRecorderProps,
     ref,
   ) => {
+    const {
+      inputDevices,
+      setPreferedDevice,
+      isStreaming,
+      startStreaming,
+      stream,
+    } = useCameras();
+
     const [maxDuration, setMaxDuration] = useState<number>(180000);
-    const [inputDevices, setInputDevices] = useState<MediaDeviceInfo[]>([]);
 
     const [recording, setRecording] = useState<boolean>(false);
     const [recorded, setRecorded] = useState<boolean>(false);
@@ -68,15 +73,6 @@ const VideoRecorder = forwardRef(
     const [saving, setSaving] = useState<boolean>(false);
     const [saved, setSaved] = useState<boolean>(false);
 
-    const [mediaStreamConstraints, setMediaStreamConstraints] =
-      useState<MediaStreamConstraints>({
-        audio: true,
-        video: {
-          facingMode: 'environment',
-          aspectRatio: VIDEO_WIDTH / VIDEO_HEIGHT,
-        },
-      });
-    const [stream, setStream] = useState<MediaStream>();
     const [mimeType, setMimeType] = useState<string>('');
 
     const [recordedChunks, setRecordedChunks] = useState<Blob[]>([]);
@@ -90,8 +86,6 @@ const VideoRecorder = forwardRef(
 
     const { uploadBlob } = useUpload(undefined, appCode);
 
-    const { device } = useBrowserInfo(navigator.userAgent);
-
     // We add one methods to handle save action from parent component
     useImperativeHandle(ref, () => ({
       save: handleSave,
@@ -99,26 +93,35 @@ const VideoRecorder = forwardRef(
 
     const { t } = useTranslation();
 
-    /**
-     * Get max duration from Conf and input devices list.
-     */
     useEffect(() => {
       initMaxDuration();
-      initInputDevices();
     }, []);
 
     /**
-     * Enable video stream and stop streaming on clean up.
+     * Get max duration from Conf.
      */
+    async function initMaxDuration() {
+      const videoConfResponse = await odeServices.video().getVideoConf();
+      setMaxDuration((videoConfResponse.maxDuration ?? 3) * 60 * 1000);
+    }
+
     useEffect(() => {
-      if (!stream) {
-        enableStream(mediaStreamConstraints);
-      }
-      return () => {
-        if (stream) {
-          stream.getTracks().forEach((track) => track.stop());
+      try {
+        if (videoRef.current) {
+          if (videoRef.current.src) {
+            window.URL.revokeObjectURL(videoRef.current.src);
+            videoRef.current.src = '';
+          }
+          if (stream) {
+            videoRef.current.srcObject = stream;
+            videoRef.current.autoplay = true;
+            videoRef.current.volume = 1;
+            videoRef.current.muted = true;
+          }
         }
-      };
+      } catch (err) {
+        console.error(err);
+      }
     }, [stream]);
 
     /**
@@ -177,72 +180,6 @@ const VideoRecorder = forwardRef(
       }
     }, [playing]);
 
-    const initMaxDuration = async () => {
-      const videoConfResponse = await odeServices.video().getVideoConf();
-      setMaxDuration((videoConfResponse.maxDuration ?? 3) * 60 * 1000);
-    };
-
-    const initInputDevices = async () => {
-      const devices = await navigator.mediaDevices.enumerateDevices();
-      const videoDevices = devices.filter(
-        (device) => device.kind === 'videoinput',
-      );
-      // console, mobile, tablet, smarttv, wearable, embedded
-      switch (device.type) {
-        case 'mobile':
-        case 'tablet': {
-          const backCamera = {
-            deviceId: 'environment',
-            label: t('video.back.camera'),
-            groupId: '',
-            kind: 'videoinput',
-          } as MediaDeviceInfo;
-          const frontCamera = {
-            deviceId: 'user',
-            label: t('video.front.camera'),
-            groupId: '',
-            kind: 'videoinput',
-          } as MediaDeviceInfo;
-
-          if (videoDevices?.length > 1) {
-            // mobile/tablet has more than 1 camera
-            setInputDevices([backCamera, frontCamera]);
-          } else {
-            // else we let the system use the only one that exists (or none)
-            setInputDevices([backCamera]);
-          }
-          break;
-        }
-        default:
-          // "Desktop" or other future types => list all cameras without distinction.
-          setInputDevices(videoDevices);
-          break;
-      }
-    };
-
-    const enableStream = async (
-      mediaStreamConstraints: MediaStreamConstraints,
-    ) => {
-      try {
-        const mediaStream: MediaStream =
-          await navigator.mediaDevices.getUserMedia(mediaStreamConstraints);
-        setStream(mediaStream);
-
-        if (videoRef.current) {
-          if (videoRef.current.src) {
-            window.URL.revokeObjectURL(videoRef.current.src);
-            videoRef.current.src = '';
-          }
-          videoRef.current.srcObject = mediaStream;
-          videoRef.current.autoplay = true;
-          videoRef.current.volume = 1;
-          videoRef.current.muted = true;
-        }
-      } catch (err) {
-        console.error(err);
-      }
-    };
-
     const handleRecord = useCallback(() => {
       setRecording(true);
 
@@ -297,7 +234,7 @@ const VideoRecorder = forwardRef(
       setRecordedTime(0);
       setRecordedChunks([]);
       setRecordedVideo(undefined);
-      enableStream(mediaStreamConstraints);
+      startStreaming();
 
       if (onRecordUpdated) {
         onRecordUpdated();
@@ -337,49 +274,22 @@ const VideoRecorder = forwardRef(
       }
     };
 
-    const handleInputDeviceChange = (option: OptionsType | string) => {
-      const selectedDevice = inputDevices.find(
-        (inputDevice) => inputDevice.label === option,
-      );
+    const handleInputDeviceChange = useCallback(
+      (option: OptionsType | string) => {
+        const selectedDevice = inputDevices.find(
+          (inputDevice) => inputDevice.label === option,
+        );
+        setPreferedDevice(selectedDevice);
 
-      let mediaStreamConstraints: MediaStreamConstraints = {};
-      if (selectedDevice?.deviceId) {
-        if (
-          selectedDevice?.deviceId === 'environment' ||
-          selectedDevice?.deviceId === 'user'
-        ) {
-          mediaStreamConstraints = {
-            audio: true,
-            video: {
-              aspectRatio: VIDEO_WIDTH / VIDEO_HEIGHT,
-              facingMode: selectedDevice?.deviceId,
-            },
-          };
-        } else {
-          mediaStreamConstraints = {
-            audio: true,
-            video: {
-              aspectRatio: VIDEO_WIDTH / VIDEO_HEIGHT,
-              deviceId: selectedDevice.deviceId,
-            },
-          };
+        // Stop any recording.
+        if (isStreaming && recorderRef.current?.state === 'recording') {
+          recorderRef.current.requestData();
+          recorderRef.current.stop();
         }
-
-        setMediaStreamConstraints(mediaStreamConstraints);
-
-        if (stream) {
-          if (recorderRef.current?.state === 'recording') {
-            recorderRef.current.requestData();
-            recorderRef.current.stop();
-          }
-          stream.getTracks().forEach((track) => track.stop());
-          setStream(undefined);
-        }
-        enableStream(mediaStreamConstraints);
-      } else {
-        console.error('Selected input device id is null');
-      }
-    };
+        startStreaming();
+      },
+      [inputDevices, isStreaming, startStreaming],
+    );
 
     /**
      * Auto-stop recording when max allowed duration is reached.
